+++ frappe_theme/public/js/custom_components/approval_request/App.vue
@@ -9,7 +9,6 @@
 import { onMounted, ref } from 'vue';
 
 const frappeContainer = ref(null);
-const ref_doctype = ref(null);
 const sva_datatable = ref(null);
 
 const props = defineProps({
@@ -19,8 +18,24 @@ const props = defineProps({
     }
 });
 
-onMounted(async () => { // ✅ made async
-    // Frappe control
+const allowedDocTypes = ref([]); // store DocTypes from SVADatatable config
+
+onMounted(async () => {
+    // Step 1: Get allowed doctypes from SVADatatable Configuration
+    if (await frappe.db.exists("SVADatatable Configuration", props.frm.doctype)) {
+        let conf_doc = await frappe.db.get_doc("SVADatatable Configuration", props.frm.doctype);
+        let found = [];
+
+        // find all child tables with link_doctype
+        for (let key in conf_doc) {
+            if (Array.isArray(conf_doc[key]) && conf_doc[key].length && conf_doc[key][0].link_doctype) {
+                found.push(...conf_doc[key].map(row => row.link_doctype).filter(Boolean));
+            }
+        }
+        allowedDocTypes.value = [...new Set(found)];
+    }
+
+    // Step 2: Create your Frappe control as before
     let moduleValue = frappe.ui.form.make_control({
         parent: $(frappeContainer.value),
         df: {
@@ -28,27 +43,50 @@ onMounted(async () => { // ✅ made async
             fieldname: 'module',
             fieldtype: 'Link',
             options: 'Workflow',
+            get_query: () => {
+                // ✅ Filter workflows to only those whose document_type
+                // is in the allowedDocTypes list for this page
+                return {
+                    filters: {
+                        document_type: ["in", allowedDocTypes.value],
+                        is_active: 1
+                    }
+                };
+            },
             onchange: async function () {
                 const newValue = moduleValue.get_value();
                 if (newValue) {
-                    let doc_type = await frappe.db.get_value('Workflow', { 'name':newValue, is_active: 1 }, 'document_type');
+                    let doc_type = await frappe.db.get_value(
+                        'Workflow',
+                        { name: newValue, is_active: 1 },
+                        'document_type'
+                    );
                     doc_type = doc_type?.message?.document_type;
-                    await showTable(doc_type);
+
+                    if (allowedDocTypes.value.includes(doc_type)) {
+                        await showTable(doc_type);
+                    } else {
+                        console.warn(`❌ ${doc_type} is not in SVADatatable Configuration list`);
+                    }
                 }
             }
         },
         render_input: true
     });
 
-    await showTable('Fund Request'); // ✅ default table
 
+    // ✅ Default table load for the DocType where this component is used
+    if (allowedDocTypes.value.includes(props.frm.doctype)) {
+        await showTable(props.frm.doctype);
+    }
 });
+
 const showTable = async (document_type) => {
     await frappe.require('sva_datatable.bundle.js');
     const frmCopy = Object.assign({}, props.frm);
     let wf_field = await frappe.db.get_value('Workflow', { document_type, is_active: 1 }, 'workflow_state_field');
-    
     wf_field = wf_field?.message?.workflow_state_field;
+
     frmCopy.sva_dt_instance = new frappe.ui.SvaDataTable({
         wrapper: sva_datatable.value,
         frm: Object.assign(frmCopy, {
@@ -88,5 +126,6 @@ const showTable = async (document_type) => {
             crud_permissions: '["read"]'
         }
     });
-}
+};
+
 </script>
